refactor: organize document asset models better
