+++ src/modules/domain/rich-text/models/document-asset/document-asset.ts
@@ -1,4 +1,4 @@
-import { type DocumentAssetSrc } from './document-asset-src';
+import { type DocumentAssetSrc } from './src';
 
 // An asset reference as it appears in a rich-text document.
 export type DocumentAsset = {
