+++ src/modules/domain/rich-text/models/document-asset/index.ts
@@ -1,0 +1,2 @@
+export * from './src';
+export * from './document-asset';
