+++ src/modules/domain/rich-text/models/index.ts
@@ -1,5 +1,4 @@
 export * from './document';
 export * from './heading';
 export * from './link';
-export * from './document-asset-src';
 export * from './document-asset';
